The start() method in Main.java is too long.
This makes the code harder to read and modify in
the future.

Refactor the start() method by breaking it into
smaller, more manageable methods.
Extract the UI setup logic and event handler
setup into separate methods.

This approach improves code readability and makes
future modifications easier.
It also ensures that each part of the code has
a clear, single responsibility.

This change makes the code more modular.
