--- src/main/java/Jerome/Launcher.java
@@ -1,13 +1,0 @@
-package Jerome;
-
-import javafx.application.Application;
-/**
- * The entry point of the JavaFX application.
- * Launches the {@link Main} class, which starts the GUI for the Jerome chatbot.
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
-
