+++ src/main/java/jerome/Launcher.java
@@ -1,0 +1,13 @@
+package jerome;
+
+import javafx.application.Application;
+/**
+ * The entry point of the JavaFX application.
+ * Launches the {@link Main} class, which starts the GUI for the Jerome chatbot.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
+
