+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 _____________________________
-Wassup, I'm Jerome.Jerome!
+Wassup, I'm jerome.jerome!
 What can I do for you?
 _____________________________
 _____________________________
